+++ pkg/apis/ml/v1alpha1/extension_status_reconciliation.go
@@ -20,21 +20,23 @@
 
 package v1alpha1
 
+import sharedApi "github.com/arangodb/kube-arangodb/pkg/apis/shared/v1"
+
 type ArangoMLExtensionStatusReconciliation struct {
-	StatefulSetChecksum string `json:"statefulSetChecksum,omitempty"`
-	ServiceChecksum     string `json:"serviceChecksum,omitempty"`
+	StatefulSet *sharedApi.Object `json:"statefulSet,omitempty"`
+	Service     *sharedApi.Object `json:"service,omitempty"`
 }
 
-func (r *ArangoMLExtensionStatusReconciliation) GetStatefulSetChecksum() string {
+func (r *ArangoMLExtensionStatusReconciliation) GetStatefulSet() *sharedApi.Object {
 	if r == nil {
-		return ""
+		return nil
 	}
-	return r.StatefulSetChecksum
+	return r.StatefulSet
 }
 
-func (r *ArangoMLExtensionStatusReconciliation) GetServiceChecksum() string {
+func (r *ArangoMLExtensionStatusReconciliation) GetService() *sharedApi.Object {
 	if r == nil {
-		return ""
+		return nil
 	}
-	return r.ServiceChecksum
+	return r.Service
 }